+++ common/src/main/java/com/coolerpromc/experienceskills/platform/services/IAttachmentHelper.java
@@ -5,6 +5,7 @@ import net.minecraft.world.entity.Entity;
 
 public interface IAttachmentHelper {
     void register();
+    <T> T get(Entity entity, AttachmentKey<T> key, T fallback);
     <T> T get(Entity entity, AttachmentKey<T> key);
     <T> void set(Entity entity, AttachmentKey<T> key, T value);
 }