+++ fabric/src/main/java/com/coolerpromc/experienceskills/platform/FabricPlatformHelper.java
@@ -3,7 +3,10 @@ package com.coolerpromc.experienceskills.platform;
 import com.coolerpromc.experienceskills.ExperienceSkillsFabric;
 import com.coolerpromc.experienceskills.api.IExperienceSkillsPlugin;
 import com.coolerpromc.experienceskills.platform.services.IPlatformHelper;
+import net.fabricmc.fabric.api.tag.convention.v2.ConventionalBlockTags;
 import net.fabricmc.loader.api.FabricLoader;
+import net.minecraft.tags.TagKey;
+import net.minecraft.world.level.block.Block;
 
 import java.nio.file.Path;
 import java.util.List;
@@ -35,5 +38,10 @@ public class FabricPlatformHelper implements IPlatformHelper {
     @Override
     public List<IExperienceSkillsPlugin> getPlugins() {
         return ExperienceSkillsFabric.PLUGINS;
+    }
+
+    @Override
+    public TagKey<Block> oreTag() {
+        return ConventionalBlockTags.ORES;
     }
 }